+++ api-controllers/FabrikBaseController.js
@@ -10,7 +10,7 @@ const errors = require('../common/errors');
 const cf = require('../broker/lib/cf');
 const bosh = require('../data-access-layer/bosh');
 const fabrik = require('../broker/lib/fabrik');
-const backupStore = require('../broker/lib/iaas').backupStore;
+const backupStore = require('../data-access-layer/iaas').backupStore;
 const catalog = require('../broker/lib/models/catalog');
 const ContinueWithNext = errors.ContinueWithNext;
 const BadRequest = errors.BadRequest;
