+++ api-controllers/ServiceFabrikAdminController.js
@@ -8,7 +8,7 @@ const errors = require('../common/errors');
 const logger = require('../common/logger');
 const config = require('../common/config');
 const NetworkSegmentIndex = require('../data-access-layer/bosh/NetworkSegmentIndex');
-const backupStore = require('../broker/lib/iaas').backupStore;
+const backupStore = require('../data-access-layer/iaas').backupStore;
 const FabrikBaseController = require('./FabrikBaseController');
 const utils = require('../broker/lib/utils');
 const fabrik = require('../broker/lib/fabrik');
