+++ api-controllers/ServiceFabrikApiController.js
@@ -5,7 +5,7 @@ const assert = require('assert');
 const Promise = require('bluebird');
 const jwt = require('../broker/lib/jwt');
 const logger = require('../common/logger');
-const backupStore = require('../broker/lib/iaas').backupStore;
+const backupStore = require('../data-access-layer/iaas').backupStore;
 const filename = backupStore.filename;
 const eventmesh = require('../eventmesh');
 const lockManager = eventmesh.lockManager;
