+++ broker/lib/fabrik/DirectorManager.js
@@ -8,7 +8,7 @@ const logger = require('../logger');
 const errors = require('../errors');
 const bosh = require('../../../data-access-layer/bosh');
 const cf = require('../cf');
-const backupStore = require('../iaas').backupStore;
+const backupStore = require('../../../data-access-layer/iaas').backupStore;
 const utils = require('../utils');
 const Agent = require('./Agent');
 const BaseManager = require('./BaseManager');
