+++ broker/lib/fabrik/OobBackupManager.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const logger = require('../logger');
-const backupStore = require('../iaas').backupStoreForOob;
+const backupStore = require('../../../data-access-layer/iaas').backupStoreForOob;
 const Agent = require('./Agent');
 const utils = require('../utils');
 const errors = require('../errors');
