+++ broker/lib/fabrik/VirtualHostRelationMapper.js
@@ -4,7 +4,7 @@ const Promise = require('bluebird');
 const logger = require('../logger');
 const bosh = require('../../../data-access-layer/bosh');
 const cloudController = require('../cf').cloudController;
-const virtualHostStore = require('../iaas').virtualHostStore;
+const virtualHostStore = require('../../../data-access-layer/iaas').virtualHostStore;
 
 class VirtualHostRelationMapper {
   constructor() {
