+++ broker/lib/jobs/BackupReaperJob.js
@@ -11,8 +11,8 @@ const errors = require('../errors');
 const ServiceInstanceNotFound = errors.ServiceInstanceNotFound;
 const NotFound = errors.NotFound;
 const cloudController = require('../cf').cloudController;
-const backupStoreForInstance = require('../iaas').backupStore;
-const backupStoreForOob = require('../iaas').backupStoreForOob;
+const backupStoreForInstance = require('../../../data-access-layer/iaas').backupStore;
+const backupStoreForOob = require('../../../data-access-layer/iaas').backupStoreForOob;
 const ScheduleManager = require('./ScheduleManager');
 const EventLogInterceptor = require('../../../common/EventLogInterceptor');
 const bosh = require('../../../data-access-layer/bosh');
