+++ broker/lib/jobs/ScheduleBackupJob.js
@@ -11,7 +11,7 @@ const utils = require('../utils');
 const retry = utils.retry;
 const ServiceInstanceNotFound = errors.ServiceInstanceNotFound;
 const cloudController = require('../cf').cloudController;
-const backupStore = require('../iaas').backupStore;
+const backupStore = require('../../../data-access-layer/iaas').backupStore;
 const ScheduleManager = require('./ScheduleManager');
 //Above reference to schedulemanager leads to the below cyclic dependency:
 // ScheduleManager -> Scheduler -> JobFabrik -> ScheduleBackupJob -> ScheduleManager
