+++ data-access-layer/iaas/AzureClient.js
@@ -2,9 +2,9 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const azureStorage = require('azure-storage');
-const logger = require('../logger');
-const errors = require('../errors');
-const utils = require('../utils');
+const logger = require('../../common/logger');
+const errors = require('../../common/errors');
+const utils = require('../../common/utils');
 const ComputeClient = require('./ComputeClient');
 const BaseCloudClient = require('./BaseCloudClient');
 const NotFound = errors.NotFound;
