+++ data-access-layer/iaas/BackupStore.js
@@ -4,12 +4,12 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const path = require('path');
 const moment = require('moment');
-const errors = require('../errors');
+const errors = require('../../common/errors');
 const NotImplementedBySubclass = errors.NotImplementedBySubclass;
-const logger = require('../logger');
-const CONST = require('../constants');
-const catalog = require('../models/catalog');
-const config = require('../config');
+const logger = require('../../common/logger');
+const CONST = require('../../common/constants');
+const catalog = require('../../broker/lib/models/catalog');
+const config = require('../../common/config');
 const BaseCloudClient = require('./BaseCloudClient');
 const NotFound = errors.NotFound;
 const Unauthorized = errors.Unauthorized;
