+++ data-access-layer/iaas/GcpClient.js
@@ -4,9 +4,9 @@ const _ = require('lodash');
 const Promise = require('bluebird');
 const GcpStorage = require('@google-cloud/storage');
 const GcpCompute = require('@google-cloud/compute');
-const logger = require('../logger');
-const errors = require('../errors');
-const utils = require('../utils');
+const logger = require('../../common/logger');
+const errors = require('../../common/errors');
+const utils = require('../../broker/lib/utils');
 const BaseCloudClient = require('./BaseCloudClient');
 const NotFound = errors.NotFound;
 const Unauthorized = errors.Unauthorized;
