+++ managers/backup-manager/BackupService.js
@@ -6,7 +6,7 @@ const config = require('../../broker/lib/config');
 const logger = require('../../common/logger');
 const errors = require('../../common/errors');
 const bosh = require('../../data-access-layer/bosh');
-const backupStore = require('../../broker/lib/iaas').backupStore;
+const backupStore = require('../../data-access-layer/iaas').backupStore;
 const utils = require('../../broker/lib/utils');
 const eventmesh = require('../../eventmesh');
 const Agent = require('../../broker/lib/fabrik/Agent');
