+++ test/test_broker/acceptance/service-fabrik-api-2.0.instances.director.spec.js
@@ -13,8 +13,9 @@ const errors = lib.errors;
 const fabrik = lib.fabrik;
 const utils = lib.utils;
 const NotFound = errors.NotFound;
-const backupStore = lib.iaas.backupStore;
-const filename = lib.iaas.backupStore.filename;
+const iaas = require('../../../data-access-layer/iaas');
+const backupStore = iaas.backupStore;
+const filename = iaas.backupStore.filename;
 
 function enableServiceFabrikV2() {
   config.enable_service_fabrik_v2 = true;
@@ -92,7 +93,7 @@ describe('service-fabrik-api-sf2.0', function () {
       before(function () {
         enableServiceFabrikV2();
         config.mongodb.provision.plan_id = 'bc158c9a-7934-401e-94ab-057082a5073f';
-        backupStore.cloudProvider = new lib.iaas.CloudProviderClient(config.backup.provider);
+        backupStore.cloudProvider = new iaas.CloudProviderClient(config.backup.provider);
         mocks.cloudProvider.auth();
         mocks.cloudProvider.getContainer(container);
         _.unset(fabrik.DirectorManager, plan_id);
@@ -118,7 +119,7 @@ describe('service-fabrik-api-sf2.0', function () {
       after(function () {
         disableServiceFabrikV2();
         timestampStub.restore();
-        backupStore.cloudProvider = lib.iaas.cloudProvider;
+        backupStore.cloudProvider = iaas.cloudProvider;
         cancelScheduleStub.restore();
         scheduleStub.restore();
         getScheduleStub.restore();
