+++ test/test_broker/acceptance/service-fabrik-api.instances.director.spec.js
@@ -13,8 +13,9 @@ const errors = lib.errors;
 const fabrik = lib.fabrik;
 const utils = lib.utils;
 const NotFound = errors.NotFound;
-const backupStore = lib.iaas.backupStore;
-const filename = lib.iaas.backupStore.filename;
+const iaas = require('../../../data-access-layer/iaas');
+const backupStore = iaas.backupStore;
+const filename = iaas.backupStore.filename;
 
 describe('service-fabrik-api', function () {
 
@@ -71,7 +72,7 @@ describe('service-fabrik-api', function () {
       before(function () {
         config.enable_service_fabrik_v2 = false;
         config.mongodb.provision.plan_id = 'bc158c9a-7934-401e-94ab-057082a5073f';
-        backupStore.cloudProvider = new lib.iaas.CloudProviderClient(config.backup.provider);
+        backupStore.cloudProvider = new iaas.CloudProviderClient(config.backup.provider);
         mocks.cloudProvider.auth();
         mocks.cloudProvider.getContainer(container);
         _.unset(fabrik.DirectorManager, plan_id);
@@ -96,7 +97,7 @@ describe('service-fabrik-api', function () {
 
       after(function () {
         timestampStub.restore();
-        backupStore.cloudProvider = lib.iaas.cloudProvider;
+        backupStore.cloudProvider = iaas.cloudProvider;
         cancelScheduleStub.restore();
         scheduleStub.restore();
         getScheduleStub.restore();
