+++ test/test_broker/iaas.GcpClient.spec.js
@@ -6,8 +6,7 @@ const Promise = require('bluebird');
 const logger = require('../../broker/lib/logger');
 const errors = require('../../broker/lib/errors');
 const utils = require('../../broker/lib/utils');
-const lib = require('../../broker/lib');
-const GcpClient = lib.iaas.GcpClient;
+const GcpClient = require('../../data-access-layer/iaas').GcpClient;
 const NotFound = errors.NotFound;
 const Forbidden = errors.Forbidden;
 const UnprocessableEntity = errors.UnprocessableEntity;
